+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `StackFormat` to format the provided stack trace as text.
+- `StackMarshalJSON` marshals the provided stack trace as JSON.
 - Support wrapping multiple errors with `errors.Join`.
   [#4](https://gitlab.com/tozd/go/errors/-/issues/4)